third_party/blink/web_tests/external/wpt/gamepad/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Gamepad interface: attribute touches
  assert_true: The prototype object must have a property "touches" expected true got false
[FAIL] GamepadHapticActuator interface: operation playEffect(GamepadHapticEffectType, optional GamepadEffectParameters)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] GamepadEvent interface object length
  assert_equals: wrong value for GamepadEvent.length expected 2 but got 1
Harness: the test ran to completion.
